implementation in cpp to avoid duplicate symbole
